Roalloc os_auth functions outside of main-server
